Update appspec.json
Hi Jed, in this (and maybe your other two Apps - please check) the link to the readme.md is not correct. Please would you update this as here suggested, then make a new release of the App here on GH and submit a new App version to MoveApps? Then it will be much easier for our reviewers to find the documentation to each app. Thanks for a quick fix! Best, Andrea